feat: add InitializeRecoilState component
