+++ client/babel.config.js
@@ -1,12 +1,12 @@
 module.exports = {
   presets: [
-    "@babel/preset-env",
-    "@babel/preset-typescript",
-    "@babel/preset-react",
+    '@babel/preset-env',
+    '@babel/preset-typescript',
+    '@babel/preset-react',
   ],
   plugins: [
     [
-      "@babel/plugin-transform-runtime",
+      '@babel/plugin-transform-runtime',
       {
         regenerator: true,
       },
